+++ project/build.properties
@@ -1,1 +1,2 @@
+# suppress inspection "UnusedProperty"
 sbt.version = 1.2.6
