+++ src/main/scala/TitTacToe.scala
@@ -1,4 +1,5 @@
 import ai._
+import ai.old.negamaxNextMove
 import game.BoardTicTacToe
 
 object TitTacToe extends App {
@@ -36,7 +37,7 @@ object TitTacToe extends App {
           }
         }
         else {
-          val (score, i, j) = negamaxNextMove(game, -1)
+          val (_, i, j) = negamaxNextMove(game, -1)
           game.playMove((i, j), computerPlayer)
         }
         playerTurn = !playerTurn
@@ -69,7 +70,7 @@ object TitTacToe extends App {
         }
 
 //        val (score, i, j) = negamaxNextMove(game, color)
-        val (score, i, j, a2 , b2) = alphaBetaNextMove(game, depth, a, b,  joshuaPlay)
+        val (_, i, j, a2 , b2) = alphaBetaNextMove(game, depth, a, b,  joshuaPlay)
         a = a2
         b = b2
         depth +=1
