+++ src/main/scala/ai/MiniMax.scala
@@ -6,41 +6,6 @@ trait MiniMax extends AiBoard {
 
   private def player(maximizing: Boolean): Byte = if (maximizing) 1 else 2
 
-  /**
-    * @deprecated used for benchmarks on FP and code reuse too
-    * @param maximizing
-    * @param depth
-    * @return
-    */
-  def solveRaw(maximizing: Boolean = true, depth: Int = 0): Int = {
-    if (gameEnded(depth)) {
-      score()
-    } else {
-      Stats.totalCalls += 1
-      if (maximizing) {
-        var value = Int.MinValue
-
-        consumeMoves() { p =>
-          playMove(p, 1)
-          value = Math.max(value, solve(!maximizing, depth + 1))
-          undoMove(p, 1)
-        }
-
-        value
-      } else {
-        var value = Int.MaxValue
-
-        consumeMoves() { p =>
-          playMove(p, 2)
-          value = Math.min(value, solve(!maximizing, depth + 1))
-          undoMove(p, 2)
-        }
-
-        value
-      }
-    }
-  }
-
   protected def mainBlock(player: Byte, depth: Int)(eval: Status => Int): Int = {
     if (gameEnded(depth)) {
       score()
