+++ src/main/scala/ai/TranspositionTable.scala
@@ -7,7 +7,7 @@ import scala.collection.mutable
 trait TranspositionTable extends BoardMN {
   val transpositions: mutable.Map[String, Transposition] = mutable.Map.empty
 
-  def hash() = board.flatten.mkString
+  def hash(): String = board.flatten.mkString
 
   def add(t: Transposition): Unit = {
     transpositions.update(hash(), t)
