+++ src/main/scala/benchmarks/MNKSolvers.scala
@@ -1,6 +1,7 @@
 package benchmarks
 
 import ai._
+import ai.old.BoardMNKwithGetBoard
 import game.BoardMNK
 
 import scala.collection.immutable.NumericRange
@@ -27,7 +28,7 @@ object MNKSolvers extends App {
       val board = new BoardMNK(m, n, k)
       val start = System.currentTimeMillis()
       Stats.totalCalls = 0
-      val score = ai.minimax(board, true)
+      val score = ai.old.minimax(board, true)
       val end = System.currentTimeMillis()
       println(s"total time: ${end - start}")
       println(s"Total calls: ${ai.Stats.totalCalls}")
@@ -56,10 +57,10 @@ object MNKSolvers extends App {
 
     {
       println("\nMiniMax trait RAW: ")
-      val board = new BoardMNK(m, n, k) with MiniMax
+      val board = new BoardMNK(m, n, k) with MiniMaxRaw
       val start = System.currentTimeMillis()
       //    Stats.totalCalls = 0
-      val score = board.solveRaw()
+      val score = board.solve()
       val end = System.currentTimeMillis()
       println(s"total time: ${end - start}")
       println(s"Total calls: ${board.Stats.totalCalls}")
@@ -75,7 +76,7 @@ object MNKSolvers extends App {
       val board = new BoardMNK(m, n, k)
       val start = System.currentTimeMillis()
       Stats.totalCalls = 0
-      val score = ai.negamax(board, 1)
+      val score = ai.old.negamax(board, 1)
       val end = System.currentTimeMillis()
       println(s"total time: ${end - start}")
       println(s"Total calls: ${ai.Stats.totalCalls}")
@@ -136,11 +137,11 @@ object MNKSolvers extends App {
 
     {
       println("\nAlpha Beta With TranspositionTableOld:")
-      val board = new BoardMNK(m, n, k) with TranspositionTableOld
+      val board = new BoardMNKwithGetBoard(m, n, k) with ai.old.TranspositionTable
       val start = System.currentTimeMillis()
       Stats.totalCalls = 0
       Stats.cacheHits = 0
-      val transposition = ai.alphaBetaWithMemOld(board, board)
+      val transposition = ai.old.alphaBetaWithMem(board, board)
       val end = System.currentTimeMillis()
       println(s"total time: ${end - start}")
       println(s"Total calls: ${ai.Stats.totalCalls}")
@@ -155,7 +156,7 @@ object MNKSolvers extends App {
 
     {
       println("\nAlpha Beta With TranspositionTable:")
-      val board = new BoardMNK(m, n, k) with TranspositionTable
+      val board = new BoardMNK(m, n, k) with ai.TranspositionTable
       val start = System.currentTimeMillis()
       Stats.totalCalls = 0
       Stats.cacheHits = 0
