+++ src/main/scala/game/BoardMNK.scala
@@ -9,202 +9,6 @@ import scala.collection.immutable.NumericRange
   * @param k number of same move of a player "in a row" (or col or diagonal)
   */
 class BoardMNK(m: Short, n: Short, k: Short) extends BoardMNKP(m, n, k, 2) {
-
-  protected def scoreRowsOld(): Int = {
-    for (i <- mIndices) {
-      val s = scoreRowOld(i)
-      if (s > 0) return s
-    }
-
-    0
-  }
-
-  /**
-    * @deprecated
-    */
-  protected def scoreRowOld(row: Short): Int = {
-    def innerLoop(i: Int, h: Byte): Int = {
-      for (j <- i + 1 until i + k) {
-        if (board(row)(j) != h) return 0
-      }
-
-      h
-    }
-
-    for (i <- NumericRange.inclusive(0, nkDiff, 1)) {
-      val h = board(row)(i)
-      if (h > 0 && innerLoop(i, h) > 0) return h
-    }
-
-    0
-  }
-
-  /**
-    * @deprecated
-    */
-  protected def scoreColOld(col: Short): Int = {
-    @tailrec
-    def cmp(i: Int, j: Int, col: Short, h: Byte): Int = {
-      if (j == k) h
-      else if (h != board(i + j)(col)) 0
-      else cmp(i, j + 1, col, h)
-    }
-
-    for (i <- mkDiffIncIndices) {
-      val h = board(i)(col)
-      if (cmp(i, 1, col, h) > 0) return h
-    }
-
-    0
-  }
-
-  /**
-    * @deprecated
-    * @return
-    */
-  protected def scoreColsOld(): Int = {
-    for (j <- nIndices) {
-      val s = scoreColOld(j)
-      if (s > 0) return s
-    }
-
-    0
-  }
-
-  /**
-    * @deprecated
-    */
-  def scoreOld(): Int = {
-    def evaluate(score: Int): Option[Int] = {
-      score match {
-        case 2 => Some(-1)
-        case 1 => Some(1)
-        case 0 => None
-        case _ => None
-      }
-    }
-
-    evaluate(scoreRowsOld())
-      .orElse(evaluate(scoreColsOld()))
-      .orElse(evaluate(scoreDiagsTL()))
-      .orElse(evaluate(scoreDiagsBR()))
-      .getOrElse(0)
-  }
-
-  /**
-    * @deprecated
-    * @return
-    */
-  protected def checkWinRowsOld(): Boolean = {
-    for {
-      i <- mIndices
-      if scoreRowOld(i) > 0
-    } return true
-
-    false
-  }
-
-  /**
-    * @deprecated
-    * @return
-    */
-  protected def checkWinColsOld(): Boolean = {
-    for {
-      j <- nIndices
-      if scoreColOld(j) > 0
-    } return true
-
-    false
-  }
-
-  protected def checkWinDiagonals(): Boolean = scoreDiagsTL() > 0 || scoreDiagsBR() > 0
-
-  /**
-    * @deprecated
-    * @return
-    */
-  protected def checkWinOld(): Boolean = {
-    checkWinRowsOld() || checkWinColsOld() || checkWinDiagonals()
-  }
-
-  // ------------ to do with lookups
-
-  /**
-    * @deprecated
-    */
-  protected def scoreDiagTL(col: Short): Int = {
-    @tailrec
-    def cmpTail(h: Byte, i: Int, start: Int): Int = {
-      if (start == k) board(i)(col)
-      else {
-        if (h != board(i + start)(col + start)) 0
-        else cmpTail(h, i, start + 1)
-      }
-    }
-
-    if (n - col >= k) {
-      for {
-        i <- mkDiffIncIndices
-        if board(i)(col) > 0
-      } {
-        val h = board(i)(col)
-        val r = cmpTail(h, i, 1)
-        if (r > 0) return r
-      }
-    }
-
-    0
-  }
-
-  /**
-    * @deprecated
-    */
-  protected def scoreDiagsTL(): Int = {
-    for (i <- nkDiffIncIndices) {
-      val s = scoreDiagTL(i)
-      if (s > 0) return s
-    }
-
-    0
-  }
-
-  /**
-    * @deprecated
-    */
-  protected def scoreDiagBR(col: Short): Int = {
-    @tailrec
-    def cmpTail(h: Byte, i: Int, start: Int, stop: Int): Int = {
-      if (start == stop) board(i)(col)
-      else if (h != board(i - start)(col + start)) 0
-      else cmpTail(h, i, start + 1, stop)
-    }
-
-    for {
-      row <- k1mIndices
-      if board(row)(col) > 0
-    } {
-      val h = board(row)(col)
-      val r = cmpTail(h, row, 1, k)
-      if (r > 0) return r
-    }
-
-    0
-  }
-
-  /**
-    * @deprecated
-    */
-  protected def scoreDiagsBR(): Int = {
-    for (j <- nkDiffIncIndices) {
-      val s = scoreDiagBR(j)
-      if (s > 0) return s
-    }
-
-    0
-  }
-
-  // -------------------------------------------
-
   override def score(): Int = {
     if (LookUps.won.getOrElse(checkWin())) lastPlayer match {
       case 2 => -1
