+++ src/test/scala/ai/old/PackageSpec.scala
@@ -1,24 +1,33 @@
-package ai
+package ai.old
 
 import game.BoardTicTacToe
 import org.scalatest.{FlatSpec, Matchers}
 
 class PackageSpec extends FlatSpec with Matchers {
-  "Tic Tac Toe alphaBeta" should "solve the game" in {
+  "Tic Tac Toe game with minimax" should "solve the game" in {
     val game = new BoardTicTacToe()
-    alphaBeta(game, maximizingPlayer = true) shouldEqual 0.0
-    alphaBetaNextMove(game, 0, Double.MinValue, Double.MaxValue, true) shouldEqual(0.0, 0, 0, 0.0, Double.MaxValue)
+
+    minimax(game, true) shouldEqual 0
+  }
+
+  "Tic Tac Toe negamax" should "solve the game" in {
+    val game = new BoardTicTacToe()
+    negamax(game, 1) shouldEqual 0
+    negamaxNextMove(game, 1) shouldEqual(0, 0, 0)
   }
 
   "Player 1 Tic Tac Toe" should "win" in {
-    val game = new BoardTicTacToe()
+    val game = new BoardTicTacToe() with withGetBoard
+    val status = new TranspositionTable {}
     game.playMove((0, 0), 1)
     game.playMove((0, 1), 1)
     game.playMove((1, 0), 2)
     game.playMove((1, 1), 1)
     game.playMove((2, 1), 2)
     game.playMove((2, 2), 2)
-    alphaBeta(game, maximizingPlayer = true) should be >= 1.0
+    negamax(game, 1) should be(1)
+    minimax(game, true) shouldEqual 1
+    alphaBetaWithMem(status, game).score should be >= 1.0
   }
 
   "Player 2 Tic Tac Toe" should "win" in {
@@ -30,6 +39,7 @@ class PackageSpec extends FlatSpec with Matchers {
     game.playMove((2, 1), 2)
     game.playMove((2, 2), 2)
     game.playMove((1, 2), 2)
-    alphaBeta(game, depth = 6, maximizingPlayer = false) should be < 0.0
+    negamax(game, -1) should be (-1) // desing cannot return -1 from the first step.
+    minimax(game, false) should be(-1)
   }
 }
